requirements.txt
g4f==0.5.7.3
discord.py==2.3.2
flask==2.3.2
python-dotenv==1.0.0
